:art: fix: 파일 디렉토리 경로 조정, domain 별로 세분화
관심, 블랙리스트, 리뷰 분리
채팅, 소셜로그인 외 테스트 완료
