+++ src/main/java/com/beteam/willu/blacklist/repository/BlacklistRepository.java
@@ -1,10 +1,10 @@
-package com.beteam.willu.user.repository;
+package com.beteam.willu.blacklist.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.beteam.willu.user.entity.Blacklist;
+import com.beteam.willu.blacklist.entity.Blacklist;
 
 public interface BlacklistRepository extends JpaRepository<Blacklist, Long> {
 	boolean existsByReceiverIdAndSenderId(Long receiver, Long sender);
