+++ src/main/java/com/beteam/willu/interest/entity/Interest.java
@@ -1,6 +1,7 @@
-package com.beteam.willu.user.entity;
+package com.beteam.willu.interest.entity;
 
 import com.beteam.willu.common.Timestamped;
+import com.beteam.willu.user.entity.User;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -16,8 +17,8 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor
-@Table(name = "blacklist")
-public class Blacklist extends Timestamped {
+@Table(name = "interests")
+public class Interest extends Timestamped {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -25,16 +26,15 @@ public class Blacklist extends Timestamped {
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "receiver_id") //받는 사람
-	//    @Column(name = "receiver_id")
 	private User receiver;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "sender_id") //보내는 사람
-	//    @Column(name = "sender_id")
 	private User sender;
 
-	public Blacklist(User receiver, User sender) {
+	public Interest(User receiver, User sender) {
 		this.receiver = receiver;
 		this.sender = sender;
 	}
+
 }
