+++ src/main/java/com/beteam/willu/user/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.beteam.willu.user.controller;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
@@ -15,13 +14,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.beteam.willu.common.ApiResponseDto;
 import com.beteam.willu.common.security.UserDetailsImpl;
-import com.beteam.willu.review.dto.ReviewRequestDto;
-import com.beteam.willu.review.dto.ReviewResponseDto;
 import com.beteam.willu.user.dto.UserRequestDto;
 import com.beteam.willu.user.dto.UserResponseDto;
 import com.beteam.willu.user.dto.UserUpdateRequestDto;
 import com.beteam.willu.user.service.UserService;
-import com.sun.jdi.request.DuplicateRequestException;
 
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -73,68 +69,4 @@ public class UserController {
 		return ResponseEntity.ok(new ApiResponseDto("회원 탈퇴 성공", 200));
 	}
 
-	//관심 유저 추가
-	@PostMapping("/interest/{id}")
-	public ResponseEntity<ApiResponseDto> addInterest(@PathVariable Long id,
-		@AuthenticationPrincipal UserDetailsImpl userDetails) {
-		try {
-			userService.addInterest(id, userDetails.getUser());
-		} catch (DuplicateRequestException e) {
-			return ResponseEntity.badRequest().body(new ApiResponseDto(e.getMessage(), HttpStatus.BAD_REQUEST.value()));
-		}
-		return ResponseEntity.status(HttpStatus.ACCEPTED)
-			.body(new ApiResponseDto("관심 유저를 추가했습니다", HttpStatus.ACCEPTED.value()));
-	}
-
-	//관심 유저 삭제
-	@DeleteMapping("/interest/{id}")
-	public ResponseEntity<ApiResponseDto> removeInterest(@PathVariable Long id,
-		@AuthenticationPrincipal UserDetailsImpl userDetails) {
-		try {
-			userService.removeInterest(id, userDetails.getUser());
-		} catch (IllegalArgumentException e) {
-			return ResponseEntity.badRequest().body(new ApiResponseDto(e.getMessage(), HttpStatus.BAD_REQUEST.value()));
-		}
-		return ResponseEntity.status(HttpStatus.ACCEPTED)
-			.body(new ApiResponseDto("관심 유저를 삭제했습니다.", HttpStatus.ACCEPTED.value()));
-	}
-
-	//차단 유저 추가
-	@PostMapping("/blacklist/{id}")
-	public ResponseEntity<ApiResponseDto> addBlacklist(@PathVariable Long id,
-		@AuthenticationPrincipal UserDetailsImpl userDetails) {
-		try {
-			userService.addBlacklist(id, userDetails.getUser());
-		} catch (DuplicateRequestException e) {
-			return ResponseEntity.badRequest().body(new ApiResponseDto(e.getMessage(), HttpStatus.BAD_REQUEST.value()));
-		}
-		return ResponseEntity.status(HttpStatus.ACCEPTED)
-			.body(new ApiResponseDto("해당 유저를 차단했습니다", HttpStatus.ACCEPTED.value()));
-	}
-
-	//차단 유저 해제
-	@DeleteMapping("/blacklist/{id}")
-	public ResponseEntity<ApiResponseDto> removeBlacklist(@PathVariable Long id,
-		@AuthenticationPrincipal UserDetailsImpl userDetails) {
-		try {
-			userService.removeBlacklist(id, userDetails.getUser());
-		} catch (IllegalArgumentException e) {
-			return ResponseEntity.badRequest().body(new ApiResponseDto(e.getMessage(), HttpStatus.BAD_REQUEST.value()));
-		}
-		return ResponseEntity.status(HttpStatus.ACCEPTED)
-			.body(new ApiResponseDto("차단을 해체했습니다.", HttpStatus.ACCEPTED.value()));
-	}
-
-	// 리뷰 작성
-	@PostMapping("/review/users/{id}")
-	public void createReview(@PathVariable Long id, @RequestBody ReviewRequestDto requestDto,
-		@AuthenticationPrincipal UserDetailsImpl userDetails) {
-		userService.createReview(id, requestDto, userDetails);
-	}
-
-	// 리뷰 데이터 조회
-	@GetMapping("/review/users/{id}")
-	public ReviewResponseDto getReviews(@PathVariable Long id) {
-		return userService.getReviews(id);
-	}
 }
